refactor(server): move tray icon out of WeaselUI
