+++ WeaselServer/WeaselServerApp.cpp
@@ -3,6 +3,7 @@
 
 WeaselServerApp::WeaselServerApp()
 	: m_handler(std::make_unique<RimeWithWeaselHandler>(&m_ui))
+	, tray_icon(m_ui)
 {
 	//m_handler.reset(new RimeWithWeaselHandler(&m_ui));
 	m_server.SetRequestHandler(m_handler.get());
@@ -22,12 +23,14 @@ int WeaselServerApp::Run()
 	win_sparkle_set_registry_path("Software\\Rime\\Weasel\\Updates");
 	win_sparkle_init();
 	m_ui.Create(m_server.GetHWnd());
+	tray_icon.Create(m_server.GetHWnd());
 	m_handler->Initialize();
 
 	int ret = m_server.Run();
 
 	m_handler->Finalize();
 	m_ui.Destroy();
+	tray_icon.RemoveIcon();
 	win_sparkle_cleanup();
 
 	return ret;
