--- WeaselUI/WeaselTrayIcon.h
@@ -1,28 +1,0 @@
-#pragma once
-#include <WeaselUI.h>
-#include <WeaselIPC.h>
-#include "SystemTraySDK.h"
-
-#define	WM_WEASEL_TRAY_NOTIFY (WEASEL_IPC_LAST_COMMAND + 100)
-
-
-class WeaselTrayIcon : public CSystemTray
-{
-public:
-	enum WeaselTrayMode {
-		INITIAL, ZHUNG, ASCII, DISABLED,
-	};
-
-	WeaselTrayIcon(weasel::UI &ui);
-
-	BOOL Create(HWND hTargetWnd);
-	void Refresh();
-
-protected:
-	virtual void CustomizeMenu(HMENU hMenu);
-
-	weasel::UIStyle &m_style;
-	weasel::Status &m_status;
-	WeaselTrayMode m_mode;
-};
-
